refactor(base): reuse basic DB config set up
